Remove asset fingerprinting for CSS asset
I was using a custom-defined 11ty shortcode to add a version to
a CSS file. This asset fingerprinting technique is commonly used
to perform cache busting.

Unfortunately, this technique is an anti-pattern for websites
deployed on Netlify, because Netlify uses its own cache
invalidation system based on the ETag HTTP response header.

My approach resulted in ALL website files to be different across
builds. This is a problem both for Netlify build times, and for
Netlify CDN propagation times.

See:

-
https://answers.netlify.com/t/support-guide-making-the-most-of-netlifys-cdn-cache/127
- https://www.netlify.com/blog/2015/09/11/instant-cache-invalidation/
- https://www.netlify.com/blog/2017/02/23/better-living-through-caching/