Remove unnecessary hidden <input> in Contact form
We need to manually add a hidden <input> tag only in
JavaScript-rendered forms. Since the Contact form of this
website is a HTML form, the hidden <input> tag is automatically
added by Netlify (after the website is built and before it is
deployed), so we don't need to manually add it.

See:
- https://docs.netlify.com/forms/setup/#work-with-javascript-rendered-forms
- https://www.seancdavis.com/blog/what-you-need-to-know-about-netlify-forms/